working iOS boot with running user-mode processes
This work is largely based on the work done by @zhuowei
that can be found at: https://worthdoingbadly.com/xnuqemu2/

This is with a little different approach trying to keep
core qemu files/modules as they are and just implement
a new iPhone machine so moved all the code to different
modules. We no longer use the base arm mahine loading
code but implement the loading code in the machine module.
We removed some of the hacks for accessing system registers
and instead emulated all the iPhone specific registers accessed.

This is a slightly different iOS version and a different iPhone.
This iPhone uses a secure monitor and doesn't use KTRR. This mostly
changes the system registers we have emulate and we need to load
the kernel twice for this device as some parsing is done by the
secure monitor at different addresses. We also need to load
the secure monitor and the boot args.

Another change is that we added properties to the device tree
so that the platform expert function no longer hangs so that
can get to executing user mode porcesses without dynamically patching
the kernel at runtime or at all.

I hope that by implementing it this way, this can be ported
more easily to other iOS versions and devices.