xnu: add support for loading custom function hooks
This add support for loading custom code at any given point
at the kernel code to override the kernel implementation of a function
and then execute the original function.

This is very useful for logging specific function execution
such as sniffing all of the mach messages sent in the system.